requirements.txt
python-dotenv
langchain
langchain-community
langchain-groq
langchain-text-splitters
chromadb
sentence-transformers
torch